requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.32
pymysql==1.1.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-dotenv==1.0.1

pytest==8.3.2
httpx==0.27.0